Organize files and namespaces
