+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Speckle.Connectors.GrasshopperShared.projitems
@@ -24,12 +24,12 @@
     <Compile Include="$(MSBuildThisFileDirectory)Components\Dev\DeconstructSpeckleParam.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Objects\CreateSpeckleObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Objects\CreateSpeckleProperties.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\ModelMenuHandler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\ProjectMenuHandler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\SearchToolStripMenuItem.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\SpeckleOperationWizard.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\VersionMenuHandler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\WorkspaceMenuHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Wizard\ModelMenuHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Wizard\ProjectMenuHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Wizard\SearchToolStripMenuItem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Wizard\VersionMenuHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Wizard\WorkspaceMenuHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Components\Operations\Wizard\SpeckleOperationWizard.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\GrasshopperMaterialUnpacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Objects\GetObjectProperties.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Components\Objects\PropertyGroupPathsSelector.cs" />
